+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/EClockingBlockInstantiation.kt
@@ -17,9 +17,9 @@
 package io.verik.compiler.ast.element.declaration.sv
 
 import io.verik.compiler.ast.common.Type
+import io.verik.compiler.ast.element.expression.common.EExpression
 import io.verik.compiler.ast.element.expression.sv.EEventControlExpression
 import io.verik.compiler.ast.property.AnnotationEntry
-import io.verik.compiler.ast.property.PortInstantiation
 import io.verik.compiler.common.TreeVisitor
 import io.verik.compiler.common.Visitor
 import io.verik.compiler.message.SourceLocation
@@ -31,12 +31,12 @@ class EClockingBlockInstantiation(
     override var type: Type,
     override var annotationEntries: List<AnnotationEntry>,
     override var documentationLines: List<String>?,
-    override val portInstantiations: List<PortInstantiation>,
+    override val valueArguments: ArrayList<EExpression>,
     var eventControlExpression: EEventControlExpression
 ) : EAbstractComponentInstantiation() {
 
     init {
-        portInstantiations.forEach { it.expression?.parent = this }
+        valueArguments.forEach { it.parent = this }
         eventControlExpression.parent = this
     }