+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/EPort.kt
@@ -18,9 +18,9 @@ package io.verik.compiler.ast.element.declaration.sv
 
 import io.verik.compiler.ast.common.Type
 import io.verik.compiler.ast.element.declaration.common.EAbstractValueParameter
+import io.verik.compiler.ast.element.expression.common.EExpression
 import io.verik.compiler.ast.property.AnnotationEntry
 import io.verik.compiler.ast.property.PortType
-import io.verik.compiler.common.TreeVisitor
 import io.verik.compiler.common.Visitor
 import io.verik.compiler.message.SourceLocation
 
@@ -32,9 +32,9 @@ class EPort(
     val portType: PortType
 ) : EAbstractValueParameter() {
 
+    override var expression: EExpression? = null
+
     override fun accept(visitor: Visitor) {
         visitor.visitPort(this)
     }
-
-    override fun acceptChildren(visitor: TreeVisitor) {}
 }