+++ verik-compiler/src/main/kotlin/io/verik/compiler/common/ElementPrinter.kt
@@ -93,7 +93,6 @@ import io.verik.compiler.ast.element.expression.sv.EWidthCastExpression
 import io.verik.compiler.ast.property.AnnotationEntry
 import io.verik.compiler.ast.property.ExpressionStringEntry
 import io.verik.compiler.ast.property.LiteralStringEntry
-import io.verik.compiler.ast.property.PortInstantiation
 import io.verik.compiler.message.Messages
 import io.verik.compiler.message.SourceLocation
 
@@ -163,7 +162,6 @@ class ElementPrinter : Visitor() {
             build(`class`.isAbstract)
             build(`class`.isObject)
             build(`class`.primaryConstructor)
-            build(`class`.superTypeCallExpression)
         }
     }
 
@@ -266,6 +264,7 @@ class ElementPrinter : Visitor() {
             build(primaryConstructor.name)
             build(primaryConstructor.type.toString())
             build(primaryConstructor.valueParameters)
+            build(primaryConstructor.superTypeCallExpression)
         }
     }
 
@@ -331,7 +330,7 @@ class ElementPrinter : Visitor() {
         build("ComponentInstantiation") {
             build(componentInstantiation.name)
             build(componentInstantiation.type.toString())
-            buildPortInstantiations(componentInstantiation.portInstantiations)
+            build(componentInstantiation.valueArguments)
         }
     }
 
@@ -339,7 +338,7 @@ class ElementPrinter : Visitor() {
         build("ModulePortInstantiation") {
             build(modulePortInstantiation.name)
             build(modulePortInstantiation.type.toString())
-            buildPortInstantiations(modulePortInstantiation.portInstantiations)
+            build(modulePortInstantiation.valueArguments)
         }
     }
 
@@ -347,7 +346,7 @@ class ElementPrinter : Visitor() {
         build("ClockingBlockInstantiation") {
             build(clockingBlockInstantiation.name)
             build(clockingBlockInstantiation.type.toString())
-            buildPortInstantiations(clockingBlockInstantiation.portInstantiations)
+            build(clockingBlockInstantiation.valueArguments)
             build(clockingBlockInstantiation.eventControlExpression)
         }
     }
@@ -356,6 +355,7 @@ class ElementPrinter : Visitor() {
         build("KtValueParameter") {
             build(valueParameter.name)
             build(valueParameter.type.toString())
+            build(valueParameter.expression)
             build(valueParameter.isPrimaryConstructorProperty)
             build(valueParameter.isMutable)
         }
@@ -365,6 +365,7 @@ class ElementPrinter : Visitor() {
         build("SvValueParameter") {
             build(valueParameter.name)
             build(valueParameter.type.toString())
+            build(valueParameter.expression)
             build(valueParameter.isInput)
         }
     }
@@ -788,16 +789,6 @@ class ElementPrinter : Visitor() {
         entries.forEach { builder(it) }
         this.builder.append("]")
         first = false
-    }
-
-    private fun buildPortInstantiations(portInstantiations: List<PortInstantiation>) {
-        buildList(portInstantiations) {
-            build("PortInstantiation") {
-                build(it.port.name)
-                build(it.port.portType.toString())
-                build(it.expression)
-            }
-        }
     }
 
     companion object {