+++ verik-compiler/src/test/kotlin/io/verik/compiler/interpret/ConstructorDesugarTransformerStageTest.kt
@@ -33,7 +33,7 @@ internal class ConstructorDesugarTransformerStageTest : BaseTest() {
                 KtClass(
                     C, C,
                     [SecondaryConstructor(C, C, BlockExpression(Unit, []), [], null)],
-                    [], 0, 0, 0, null, null
+                    [], 0, 0, 0, null
                 )
             """.trimIndent()
         ) { it.findDeclaration("C") }
@@ -61,8 +61,8 @@ internal class ConstructorDesugarTransformerStageTest : BaseTest() {
                 KtClass(
                     C, C, [
                         Property(x, Int, null, 0),
-                        SecondaryConstructor(C, C, $blockExpression, [KtValueParameter(x, Int, 0, 0)], null)
-                    ], [], 0, 0, 0, null, null
+                        SecondaryConstructor(C, C, $blockExpression, [KtValueParameter(*)], null)
+                    ], [], 0, 0, 0, null
                 )
             """.trimIndent()
         ) { it.findDeclaration("C") }