+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Rename `onr` to `oni`.
 - Support secondary constructors.
 - Replace `ENullExpression` with `ENothingExpression`.
+- Support optional value parameters.
 
 ## [0.1.13]
 ### Importer